fix: api 위치 API Controller로 수정
